continued refactoring.
