+++ src/reducers/index.js
@@ -1,5 +1,5 @@
 import app from './app';
 import game from './game';
-import { combineReducers } from 'redux'
+import {combineReducers} from 'redux';
 
 export default combineReducers({app, game});
